Move Router into a separate file
For readability purposes, and to not overwhelm the developer when they pull this template down as part of their project, I've moved the `Router` class into `router.js`. `index.js` now imports the class from that file - the rest of `index.js` hasn't been changed.
